Cleanup datastore package
* Remove fetch api
* Move it to projector, since it is used there
