+++ cmd/datastore/example-data.json.go
@@ -800,7 +800,7 @@ var exampleData = map[string]json.RawMessage{
                 5
             ]`),
 	"meeting/1/assignment_poll_default_method":            []byte(`"votes"`),
-	"meeting/1/assignment_poll_default_type":              []byte(`"nominal"`),
+	"meeting/1/assignment_poll_default_type":              []byte(`"pseudoanonymous"`),
 	"meeting/1/assignment_poll_sort_poll_result_by_votes": []byte(`true`),
 	"meeting/1/assignments_export_preamble":               []byte(`""`),
 	"meeting/1/assignments_export_title":                  []byte(`"Elections"`),
@@ -950,7 +950,7 @@ var exampleData = map[string]json.RawMessage{
                 2,
                 3
             ]`),
-	"meeting/1/motion_poll_default_type": []byte(`"analog"`),
+	"meeting/1/motion_poll_default_type": []byte(`"pseudoanonymous"`),
 	"meeting/1/motion_state_ids": []byte(`[
                 1,
                 2,
@@ -1011,7 +1011,7 @@ var exampleData = map[string]json.RawMessage{
 	"meeting/1/motions_statute_recommendations_by":            []byte(`"Statute ABK"`),
 	"meeting/1/motions_statutes_enabled":                      []byte(`true`),
 	"meeting/1/motions_supporters_min_amount":                 []byte(`1`),
-	"meeting/1/name": []byte(`"OpenSlides - Die Veranstaltung (Teil II)"`),
+	"meeting/1/name": []byte(`"OpenSlides Demo"`),
 	"meeting/1/option_ids": []byte(`[
                 1,
                 2,
@@ -1111,13 +1111,13 @@ var exampleData = map[string]json.RawMessage{
                 3
             ]`),
 	"meeting/1/users_allow_self_set_present": []byte(`true`),
-	"meeting/1/users_email_body":             []byte(`"Dear {name},\n\nthis is your OpenSlides login for the event {event_name}:\n\n    {url}\n    username: {username}\n    password: {password}\n\nThis email was generated automatically."`),
+	"meeting/1/users_email_body":             []byte(`"Dear {name},\n\nthis is your personal OpenSlides login:\n\n{url}\nUsername: {username}\nPassword: {password}\n\n\nThis email was generated automatically."`),
 	"meeting/1/users_email_replyto":          []byte(`""`),
 	"meeting/1/users_email_sender":           []byte(`"Openslides"`),
-	"meeting/1/users_email_subject":          []byte(`"Your login for {event_name}"`),
+	"meeting/1/users_email_subject":          []byte(`"OpenSlides access data"`),
 	"meeting/1/users_enable_presence_view":   []byte(`true`),
 	"meeting/1/users_enable_vote_weight":     []byte(`true`),
-	"meeting/1/users_pdf_url":                []byte(`"http://example.com:8000"`),
+	"meeting/1/users_pdf_url":                []byte(`"https://example.com"`),
 	"meeting/1/users_pdf_welcometext":        []byte(`"[Place for your welcome and help text.]"`),
 	"meeting/1/users_pdf_welcometitle":       []byte(`"Welcome to OpenSlides"`),
 	"meeting/1/users_pdf_wlan_encryption":    []byte(`""`),
@@ -1962,8 +1962,8 @@ var exampleData = map[string]json.RawMessage{
             ]`),
 	"organization/1/theme_id":            []byte(`1`),
 	"organization/1/theme_ids":           []byte(`[1, 2, 3]`),
-	"organization/1/url":                 []byte(`"http://example.com:8000"`),
-	"organization/1/users_email_body":    []byte(`"Dear {name},\n\nthis is your personal OpenSlides login:\n{url}\nusername: {username}\npassword: {password}\n\n\nThis email was generated automatically.\n"`),
+	"organization/1/url":                 []byte(`"https://example.com"`),
+	"organization/1/users_email_body":    []byte(`"Dear {name},\n\nthis is your personal OpenSlides login:\n\n{url}\nUsername: {username}\nPassword: {password}\n\n\nThis email was generated automatically."`),
 	"organization/1/users_email_replyto": []byte(`""`),
 	"organization/1/users_email_sender":  []byte(`"OpenSlides"`),
 	"organization/1/users_email_subject": []byte(`"OpenSlides access data"`),
