+++ internal/projector/datastore/alias.go
@@ -1,0 +1,9 @@
+package datastore
+
+import "github.com/OpenSlides/openslides-autoupdate-service/pkg/datastore"
+
+// NewRecorder from datastore.NewRecorder
+var NewRecorder = datastore.NewRecorder
+
+// DoesNotExistError is a type alias from datastore.DoesNotExistError
+type DoesNotExistError = datastore.DoesNotExistError
