+++ internal/projector/slide/assignment.go
@@ -7,7 +7,7 @@ import (
 	"sort"
 
 	"github.com/OpenSlides/openslides-autoupdate-service/internal/projector"
-	"github.com/OpenSlides/openslides-autoupdate-service/pkg/datastore"
+	"github.com/OpenSlides/openslides-autoupdate-service/internal/projector/datastore"
 )
 
 type dbAssignment struct {
